requirements.txt
Flask
requests
beautifulsoup4
pandas
openpyxl
python-dotenv
flask-cors 
